+++ StochasticPIM/ASC_adder.h
@@ -1,4 +1,4 @@
 #ifndef ASC_ADDER_H_
 #define ASC_ADDER_H_
-float** ASC_adder(float** x1, int n, float schThrPerc, float K, float feedB, float schOutLevel, float timeStep, float samples);
+float** ASC_adder(float** x1, int n, float schThrPerc, float K, float feedB, float schOutLevel, float timeStep, int samples);
 #endif
